+++ NEWS.md
@@ -25,6 +25,8 @@ implementation, and was not being used in any actual calculation.
 computed for Acute risk.
 * Improvements in the UI of the Rice Field Management System's editor, 
 addressing #478.
+* Improvements in the appearance of the "Allocation Rules" section (which is
+no longer a dismissable sidebar, but a fixed card). Also addresses #482.
 * Implemented chemical selection in risk output plots (#476).
 * The Y-axis of Risk Quotient plots now reads `PEC/PNEC`.
 * Various improvements in the matching of parameter names between documentation